Design Beta: Clean Up Welcome Header Slash Animation
The Welcome Header Slash animation worked but the timing was off
and it could take a while before the animation would reset again.

By using the `transitionend` to switch between adding the
"transformation" and a setting the "transitionDuration" to the
random values we want the animation to take, timing is more
accurate.

We also adusted the transform X so that on smaller screens it
doesn't move as far, keeping the timing more accurate to when
it "moves off screen."

The timings still not perfect b/c we only really adjust for 3 sizes
but it works and isn't a big issue.

Part of #108