Design Beta: Update Welcome Header Images
We'd like to support a variety of floating images in the welcome
header but don't really want them all in our "static" git repository.

This adds some basic logic to pull images on the live site from
our google cloud storage instead of always expecting it to be
in the static directory.

This also updates the logic to display the header based off of the
page type b/c we have all the static assets in dev and want to
pull values from the cloud on the live site.

Part of #116