Design Beta: Add Welcome Background Animation
It was brought up that a bird flying in the background would look
cool and I thought it would be a fun challenge.

Turned out to be a bit more work that I expected, getting the background
right required another layer of html and some fancy css.

I also spent a lot of time playing around with the animation behavior,
making it move towards the edges of the screen and adding some effects
when the user clicks on the portrait.

NOTE: This changes quite a lot of text and sizes but actually kinda
fixes the page layout.

Closes #108