Welcome Header: Move to home-page.less
There's no reason for the header specific to the home-page to be
in the general css file.

This moves the welcome css into the more specific file.

It also adds another animation to the color changes so that they
will faze into the color instead of just be an immediate change.

Part of #113