Design Beta: Use Polygon for Welcome Header Slash
The clip-path/polygon shapes are fun and we can play around with
colors if we use them. They also look clean at all resolutions and
don't require images so really lots of wins.

It's pretty simple to design them with:
https://bennettfeely.com/clippy/

This converts the `orange_lines.png` image to a `clip: poly()` css
rule and adds some "wild" features that lets it change colors when
the user clicks the profile image.

Part of #113