Design Beta: Clean up X hover behavior
There was some intention to let the user interact with the floating
X's on the Welcome Page Header but due to the common "transform delay"
it was difficult to see thing were happening.

The silly hack is to switch the long transition delay (which we want
for the floating effect) in the few cases we want to change the
color quickly.

It does effect the floating X (it will move faster) but that kinda
works imo and looks like it runs away.

Sometimes the "quick" effect sticks but it's fun b/c it flashes colors
until the next reset, then it slows down again.

The trigger is a bit odd to do it too (I'm not sure quite what does it)
but it's a fun interaction to hover and click on so a success afaik.

Part of #112