Design Beta: Improve Welcome Header X hover interaction
The X's will change color and then store that color and try to
revert to it when the user hovers over it.

This can cause some odd iteractions b/c we also want to change
color at arbitrary times and have some logic for then a color
change finishes.

The whole point is to have a fun interactive header so I kept the
odd effects around. Getting in to the odd effect has to do with
clicking several times and hovering after time has passed.

Getting out was kinda difficult but this change makes it so when
you hover over the X again it goes "normal."

Also if it goes odd for whatever reason, it'll mellow out at some
point.

Part of #113