Design Beta: Make Welcome Header Text Scale
All the Welcome Header text sizes were defined in px but the
cool kids use em to so that it scales with the browser font sizes.

The main problem is that the current header logic has a set height
and anything overflows it.

Because I setup all the background layers, we can switch some things
around and have the "text" info lay define the header height.

Then when text size increases, the text changes the header height.

Switching the font sizes to em was pretty basic, the only odd bit
was having to switch some html elements to match so the sizing
would be the same for the icons and the text.

For fun I also updated the slash to always be the header height, so
it will grow with the container. Also when the user clicks on the
portrait all the sizes increase slightly. If they click enough
it'll show off the cool header expanding thing.

Part of #113