Design Beta: Clean up section CSS and add Welcome Content
Now that we've got a cool header, we'd like the page to start actually
having decent content too.

This updates the welcome page to have better content and updates a
lot of the page css to have proper margins and font sizes. It also
adds the Roboto font and uses that in a couple places.

Part of #107, #113