Design Beta: Add animation to Robots Header
We had the idea of making the robots kinda bounce around to make it
look like they're "moving" across the moon.

I thought it would be fun if the moon scrolled faster and they bounced
more if you clicked on them so I added that.

Nothing too fancy here, just adding the html, css and js.

Kinda bugged in that you can only click on certain part of the robot
b/c it's obscured by the text elements but idc much and can fix later
when we refine stuff it it bothers me too much.

Closes #110