Design Beta: Support HQ Robot Images on Robots Page
The image we used for the Robots page was a single, low resolution
image so that we could figure out the animation and general css.

Now that that's working, we'd like to use the higher quality images.

This does some abstracting of the silly HeaderElement, which basiclly
allows us to hack around "static" vs "google cloud" images.

Part of #112