Design Beta: Add Nav images. Minor css & content fixes
This adds the Nav images to be more relevant to the pages they link
to.

Yes the welcome page spells Chris wrong. I'll probably get higher
quality images for the live site at some point but I think it's
funny so I'll commit the dev images with the wrong spelling.

This also fixes the date for Tractor Hacking (2017 not 2007), splits
a run-on sentence and adds `pointer-events: none` to some text so
the events behind it aren't hidden.

Part of #113