Design Beta: Add Work Header background shape
Currently the Work Header is an image and we don't really like that.

We'd prefer not to use images and this gave me an excuse to play
with SVG shapes more.

This does an approximate shape to what we in the design. It's close
enough and looks good so we'll go with it.

This adds the templates & css to do the thing.

Part of #110