Design Beta: Animate the Work Header background shape
I really wanted the background shape to transition between a lot
of different colors, but for a variety of reasons that is difficult
to do.

This is a bit of a hack, just making one large grandiant and panning
left and right on it to get a gentle "change in colors" but I think
it looks pretty good and if I want to dig into it I can probably get
a lot fancier by repeating colors and making the grandiant larger.

But for now this is pretty good so we'll go with it.

Part of #110