Design Beta: Add color change to Work Header Shape
The work page header has a shape in the background that slowly
changes colors over a gradient.

I thought it would be cool if the colors would start to randomly
change over time, so I added that.

Now the first "loop" will keep the initial colors but the end
of the second "loop" will end up on different colors.

Each loop going forward will shift into new random colors through
the gradiant.

It's a bit of a silly hack to blow up the background image to a
huge size and switch out the gradiants but it seems to work pretty
well and doesn't require too much logic.

Part of #110