Design Beta: Add work video to header
I tried a bunch of things to not make the animation css instead of
a video but in the end it was too much to handle. The assets weren't
quite right & the movement animations would sorta look ok but nothing
as crisp or nice as the video.

So we'll just use the video, it's still pretty small and looks great.

Part of #109