Design Beta: Add work header text and action html
The work header currently just has a title and a video. The design
has a little bit more, my name and some of the article list tags you
can interact with.

This adds the basic html and css to get things looking more like the
mockup.

There's a big hoop we have to jump through, the video can't have a
transparent background (because reasons). We have to make sure where
it's positioned will always be black.

This gets handled pretty easily in the larger view but on mobile the
shape doesn't really work for 1/2 the page and it also doesn't reshape
well.

There's a slight solution that kinda works but having a mobile shape
is probably easiest.

NOTE: Ammended missing files

Part of #109