Design Beta: Add actions to work header buttons
The work header had some buttons with tags but they didn't do anything
when you clicked on them.

We want them to act like the tag filter on the Article List below.

Thankfully the code was written with this in mind so it just took a couple
minor changes:
 * Data additions to the header tags
 * JS changes to look for tag filters on the entire page
 * Also added a jump when the header tag was clicked to the Article List

NOTE: The tags are hardcoded b/c I didn't want to make the logic to figure
out what tags go where.

Part of #109