Design Beta: Add resume vanity url
I wanted `/resume` to show my resume but the actual file is hosted
elsewhere as a `pdf`.

Displaying it through PHP and not the native browser is a ton of work
but redirects are a thing.

We only have a single use case and there's no need to make fancy so this
is a KISS implemtation.

The only funky stuff is adding the `résumé` characters which requires url
encoding. I thought it'd be fun b/c it's jibberish if you look at the html.

closes #120