Design Beta: Add preview to resume vanity url
I'd like /resume to be a link I send around.

It currently is just a redirect but I'd like to to also have a cool
"preview" when I send in in messanger.

It's not too difficult, just needs some meta tags hacked in.

At this point we're KISS so we won't worry about non-redirect pages.

Part of #120